fix(client)!: change admin api route path
